Moved restrictions to its own file.
