+++ tests/tests/custom_authorities_tests.cpp
@@ -25,7 +25,7 @@
 #include <boost/test/unit_test.hpp>
 
 #include <graphene/chain/custom_authority_object.hpp>
-#include <graphene/chain/protocol/restriction.hpp>
+#include <graphene/chain/protocol/restrictions.hpp>
 #include <graphene/chain/protocol/operations.hpp>
 #include <graphene/chain/account_object.hpp>
 #include <graphene/chain/custom_authorities_utils.hpp>
@@ -89,205 +89,6 @@ BOOST_AUTO_TEST_CASE( validation_passes_when_now_is_in_valid_period )
 BOOST_AUTO_TEST_SUITE_END()
 
 BOOST_AUTO_TEST_SUITE( custom_authority_restrictions )
-
-template <typename Action>
-struct base_restriction
-{
-    generic_member value;
-    std::string argument;
-    
-    template <typename Operation>
-    bool validate( const Operation& op ) const
-    {
-        try
-        {
-            member_visitor<Operation, Action> visitor(argument, Action(value), op);
-            fc::reflector<Operation>::visit(visitor);
-            return true;
-        }
-        catch (...)
-        {
-            return false;
-        }
-    }
-};
-
-template <typename Action>
-struct base_list_restriction
-{
-    std::vector<generic_member> values;
-    std::string argument;
-    
-    template <typename Operation>
-    bool validate( const Operation& op ) const
-    {
-        try
-        {
-            member_visitor<Operation, Action> visitor(argument, Action(values), op);
-            fc::reflector<Operation>::visit(visitor);
-            return true;
-        }
-        catch (const fc::exception& e)
-        {
-            return false;
-        }
-    }
-};
-
-class equal
-{
-public:
-    equal(const generic_member& value)
-    : m_value(value)
-    {}
-    
-    template <class T>
-    void operator () (const T& member) const
-    {
-        FC_ASSERT(is_equal(get<T>(m_value), member));
-    }
-    
-private:
-    generic_member m_value;
-};
-
-class not_equal
-{
-public:
-    not_equal(const generic_member& value)
-    : m_value(value)
-    {}
-    
-    template <class T>
-    void operator () (const T& member) const
-    {
-        FC_ASSERT(!is_equal(get<T>(m_value), member));
-    }
-    
-private:
-    generic_member m_value;
-};
-
-class any_of
-{
-public:
-    any_of(const std::vector<generic_member>& values)
-    : m_values(values)
-    {}
-    
-    template <class T>
-    void operator () (const T& member) const
-    {
-        for (const generic_member& value: m_values)
-        {
-            if (is_equal(get<T>(value), member))
-            {
-                return;
-            }
-        }
-        
-        FC_THROW("Argument was not present in the list.");
-    }
-    
-private:
-    std::vector<generic_member> m_values;
-};
-
-class none_of
-{
-public:
-    none_of(const std::vector<generic_member>& values)
-    : m_values(values)
-    {}
-    
-    template <class T>
-    void operator () (const T& member) const
-    {
-        for (const generic_member& value: m_values)
-        {
-            if (is_equal(get<T>(value), member))
-            {
-                FC_THROW("Operation member is present in the list.");
-            }
-        }
-    }
-    
-private:
-    std::vector<generic_member> m_values;
-};
-
-class contains_all
-{
-public:
-    contains_all(const std::vector<generic_member>& values)
-    : m_values(values)
-    {}
-    
-    template <class T>
-    void operator () (const T&) const
-    {
-        FC_ASSERT("Not list type come.");
-    }
-    
-    template <class T>
-    void operator () (const flat_set<T>& list) const
-    {
-        for (const generic_member& value: m_values)
-        {
-            bool contains = false;
-            for (const auto& item: list)
-            {
-                contains |= (item == value.get<T>());
-            }
-            
-            FC_ASSERT(contains);
-        }
-    }
-    
-private:
-    std::vector<generic_member> m_values;
-};
-
-class contains_none
-{
-public:
-    contains_none(const std::vector<generic_member>& values)
-    : m_values(values)
-    {}
-    
-    template <class T>
-    void operator () (const T&) const
-    {
-        FC_ASSERT("Not list type come.");
-    }
-    
-    template <class T>
-    void operator () (const flat_set<T>& list) const
-    {
-        for (const generic_member& value: m_values)
-        {
-            for (const auto& item: list)
-            {
-                if (is_equal(item, get<T>(value)))
-                {
-                    FC_THROW("Should not contain any of same items.");
-                }
-            }
-        }
-    }
-    
-private:
-    std::vector<generic_member> m_values;
-};
-
-typedef base_restriction<equal>              eq_restriction;
-typedef base_restriction<not_equal>          neq_restriction;
-typedef base_list_restriction<any_of>        any_restriction;
-typedef base_list_restriction<none_of>       none_restriction;
-typedef base_list_restriction<contains_all>  contains_all_restriction;
-typedef base_list_restriction<contains_none> contains_none_restriction;
-
-typedef fc::static_variant<eq_restriction, neq_restriction, any_restriction, none_restriction, contains_all_restriction, contains_none_restriction> restriction;
 
 BOOST_AUTO_TEST_CASE( validation_passes_for_eq_restriction_when_assets_are_equal )
 {
